Changed playlist logic
In the previous commit we moved to an architecture where surface would
make a get playlist call on load that retrieved the user’s current
playlist instead of returning a new playlist. New playlist generation
requests were only made once the current playlist had reached its end.
The motivation was to allow the user to resume playback where they had
left off in a previous session. That logic created the following issues:

1. If the user left surface while the last video in the current
playlist was playing, they would receive a new playlist on their next
visit and therefore not be able to resume where they left off. This
happens because surface automatically requests a new playlist to be
generated when there are no more videos in the current queue (i.e. the
last video in the playlist is playing). The user’s current playlist is
updated at the server, causing state loss.

2. If the user reloads the page while a new playlist is being
generated, the user initially get's zero videos from the server because
their remote playlist is still being generated. Surface catches this
exception and immediately requests another playlist, which results in
10 videos being added to the queue. Complete and skip operations on the
first 5 videos do not work since the remote playlist does not have
these videos in it’s queue.

To fix these issues, we moved back to the old architecture where
surface always requests the generation of new playlists, with the
following changes:
1) On first load, only the last_played video is inserted into the
playlist. The player displays immediately and therefore does not have
to wait for the server to respond before becoming active. When the
player pulls the file from the queue, surface automatically requests a
new playlist from the server, thereby allowing the new playlist to be
generated asynchronously.
2) Skip button automatically becomes inactive while the first new
playlist is loading.